@@ -1,4 +1,5 @@
-import { useRouter } from "next/router"
+"use client"
+import { useParams, useSearchParams } from "next/navigation"
 import React, { useRef } from "react"
 import Image from "react-bootstrap/Image"
 
@@ -6,7 +7,7 @@ import { useReactToPrint } from "react-to-print"
 
 import { gql } from "@apollo/client"
 
-import { ACTIONS, ACTION_TYPE } from "../../pages/_reducer"
+import { ACTIONS, ACTION_TYPE } from "../../app/_reducer"
 
 import { useLnInvoicePaymentStatusSubscription } from "../../lib/graphql/generated"
 
@@ -33,8 +34,10 @@ gql`
 `
 
 function PaymentOutcome({ paymentRequest, paymentAmount, dispatch }: Props) {
-  const router = useRouter()
-  const { amount, sats, username, memo } = router.query
+  const searchParams = useSearchParams()
+  const query = searchParams ? Object.fromEntries(searchParams.entries()) : {}
+  const { username } = useParams()
+  const { amount, sats, memo } = query
   const componentRef = useRef<HTMLDivElement | null>(null)
 
   const printReceipt = useReactToPrint({
