@@ -1,8 +1,9 @@
-import { useRouter } from "next/router"
 import React from "react"
 import { Modal } from "react-bootstrap"
 
-import { ACTIONS, ACTION_TYPE } from "../../pages/_reducer"
+import { useSearchParams, useParams } from "next/navigation"
+
+import { ACTIONS, ACTION_TYPE } from "../../app/_reducer"
 
 import styles from "./memo.module.css"
 
@@ -12,35 +13,32 @@ interface Props {
 }
 
 const Memo = ({ state, dispatch }: Props) => {
-  const router = useRouter()
-  const { username, amount, sats, unit, memo, display } = router.query
+  const searchParams = useSearchParams()
+  const { username } = useParams()
+  const query = searchParams ? Object.fromEntries(searchParams.entries()) : {}
+
+  const { amount, sats, unit, memo, display } = query
   const [openModal, setOpenModal] = React.useState<boolean>(false)
   const [note, setNote] = React.useState<string>(memo?.toString() || "")
 
   const handleSetMemo = () => {
     if (unit === "SAT" || unit === "CENT") {
-      router.push(
-        {
-          pathname: `${username}`,
-          query: {
-            amount: amount,
-            sats: sats,
-            unit: unit,
-            memo: note,
-            display,
-          },
-        },
-        undefined,
-        { shallow: true },
+      window.history.pushState(
+        {},
+        "",
+        `${username}?${new URLSearchParams({
+          amount: amount,
+          sats: sats,
+          unit: unit,
+          memo: note,
+          display,
+        }).toString()}`,
       )
     } else {
-      router.push(
-        {
-          pathname: `${username}`,
-          query: { memo: note, display },
-        },
-        undefined,
-        { shallow: true },
+      window.history.pushState(
+        {},
+        "",
+        `${username}?${new URLSearchParams({ memo: note, display }).toString()}`,
       )
     }
     handleClose()
